moved Difficulty to GeoCode the subsystem
