+++ backend/src/main/java/tech/geocodeapp/geocode/geocode/request/GetGeoCodesByDifficultyRequest.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.validation.annotation.Validated;
 import javax.validation.Valid;
 
-import tech.geocodeapp.geocode.collectable.model.Difficulty;
+import tech.geocodeapp.geocode.geocode.model.Difficulty;
 
 import java.util.Objects;
