+++ backend/src/main/java/tech/geocodeapp/geocode/trackable/request/GetTrackablesRequest.java
@@ -2,7 +2,7 @@ package tech.geocodeapp.geocode.trackable.request;
 
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import tech.geocodeapp.geocode.collectable.model.Difficulty;
+import tech.geocodeapp.geocode.geocode.model.Difficulty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.ArrayList;
 import java.util.List;
